move modules in nenecessary packages and update README-file
