+++ README.md
@@ -1,6 +1,34 @@
 # проект виджет банковских операций
 
-## проект виджет банкоских операций - виджет, который показывает несколько последних успешных банковских операций клиента
+проект виджет банкоских операций - виджет, который показывает несколько последних успешных банковских операций клиента
+
+## Структура проекта
+
+```commandline
+
+|
+├──scr/                            # исходный код
+|   ├──__init__.py                 # инициализация основного пакета
+    ├──generators.py               # модуль с генераторами
+|   ├──masks.py                    # модуль с функциями маскировки счетов и номеров карт
+|   ├──processing.py               # модуль для обработки данных
+|   ├──widget.py                   # модуль для основной функциональности
+|
+├──tests/                          # пакет с тестами
+|    ├──__init__.py                # инициализация пакета с тестами
+|    ├──conftest.py                # модуль с фикстурами
+|    ├──test_generators.py         # модуль с тестами генераторов
+|    ├──test_masks.py              # модуль с тестами маскировки
+|    ├──test_processing.py         # модуль с тестами обработки данных
+|    ├──test_widget.py             # модуль с тестами для основной функциональности
+|
+├──.flake8                         # настройки для проверки качества кода
+├──.gitignore                      # игнорируемые файлы для git
+├──poetry.lock                     # файл блокировки зависимостей
+├──pyproject.toml                  # список установленных в проект зависимостей
+├──README.md                       # основная документация проекта
+    
+```
 
 ## Установка
 
@@ -13,7 +41,7 @@ git clone https://github.com/grmarcw/widget-client-banking-transactions.git
 poetry install
 ```
 ## Тестирование
-1. Типы тестов
+1. Используемые типы тестов
 
 В проекте используются Юнит-тесты(тестирование отдельных функций)
